# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.Wish:
    properties:
      approved:
        type: boolean
      authorName:
        type: string
      content:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  service.WishCreateRequest:
    properties:
      authorName:
        maxLength: 100
        type: string
      content:
        maxLength: 500
        type: string
      email:
        maxLength: 150
        type: string
    required:
    - authorName
    - content
    type: object
  service.WishListResponse:
    properties:
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
      wishes:
        items:
          $ref: '#/definitions/service.Wish'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/wishes:
    get:
      consumes:
      - application/json
      description: 获取祝福列表，支持分页查询。未认证用户只能查看已审核的祝福，已认证用户可以查看所有祝福。
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 是否只获取已审核的祝福，仅对已认证用户有效
        in: query
        name: approved
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.WishListResponse'
              type: object
      summary: 获取祝福列表
      tags:
      - wishes
    post:
      consumes:
      - application/json
      description: 创建一个新的祝福
      parameters:
      - description: 祝福信息
        in: body
        name: blessing
        required: true
        schema:
          $ref: '#/definitions/service.WishCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.Wish'
              type: object
      summary: 创建祝福
      tags:
      - wishes
  /api/v1/wishes/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的祝福
      parameters:
      - description: 祝福 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 删除祝福
      tags:
      - wishes
  /api/v1/wishes/{id}/approve:
    put:
      consumes:
      - application/json
      description: 批准指定的祝福，使其可以公开显示
      parameters:
      - description: 祝福 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 批准祝福
      tags:
      - wishes
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: Grants read and write access to administrative information
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/v1/user/token
    type: oauth2
swagger: "2.0"